Backend/requirements.txt
fastapi[standard]
sqlmodel
pymysql
passlib[bcrypt]
argon2-cffi
pydantic-settings
python-jose[cryptography]
openai
python-dotenv